compose: Overhaul fetch-osie script
Another commit where more than one thing is happening but I don't really
feel like splitting into many. Changes are as follows:

Use `set -euxo pipefail` instead of just `set -xo pipefail`:
-o pipefail w/o -e is basically useless since it won't stop execution, just
stops the command early, and -u adds to the safety.

Get rid of code for tinkerbell/osie:
This simplifies almost all of the file since hook is pretty much ready to go
as is.

Get rid of the whole indexing stuff:
No need for indices when the file names are unique anyway. If there's no indices
the filenames alone work and we don't need to figure out how to "build" up a
filename for the downloaded file.

I noticed that the /destination parameter is never used so dropped it.

All in all, a *very* nice simplification/cleanup of this code.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>