compose: Clean up volume rw/ro settings
Lets get rid of explicit `rw` as thats default anyway and its confusing to have
some `:rw` and some not specified, but both are `rw`. I've gone and made all
source script mounts be `ro` so that none of the containers modify any of the
files in git. I've also made the auth and certs volumes ro for the services
that don't need to modify any data.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>